chore: sever terminal_view <-> tasks_ui dependency
