+++ crates/task/Cargo.toml
@@ -21,6 +21,6 @@ serde_json_lenient.workspace = true
 sha2.workspace = true
 shellexpand.workspace = true
 util.workspace = true
-
+zed_actions.workspace = true
 [dev-dependencies]
 gpui = { workspace = true, features = ["test-support"] }
