+++ crates/tasks_ui/Cargo.toml
@@ -25,7 +25,7 @@ ui.workspace = true
 util.workspace = true
 workspace.workspace = true
 language.workspace = true
-
+zed_actions.workspace = true
 
 [dev-dependencies]
 editor = { workspace = true, features = ["test-support"] }
